Написал express сервер, разделил index.js на модули, изменил структуру проекта
