+++ public/css/common.css
@@ -87,3 +87,119 @@
 .sidebar-button:hover {
     cursor: pointer;
 }
+
+.redirect:hover {
+    cursor: pointer;
+}
+
+
+
+.main-page {
+    background: linear-gradient(to bottom, #141414, #252525) fixed;
+    display: grid;
+    grid-template-columns: 92px auto;
+    grid-template-rows: 112px auto;
+    grid-template-areas:
+            "header header"
+            "sidebar page";
+    gap: 0;
+}
+
+.navbar {
+    position: fixed;
+    background: linear-gradient(to bottom, #141414, #252525) fixed;
+    display: flex;
+    width: 100%;
+    height: 112px;
+    padding: 32px 0;
+    align-items: center;
+    justify-content: space-between;
+    z-index: 100;
+}
+
+.navbar-logo {
+    align-self: center;
+    margin: 0 16px;
+    float: left;
+}
+
+.search-form {
+    flex-grow: 1;
+    height: 48px;
+    margin-right: 20px;
+    padding: 10px;
+    border-radius: 20px;
+    border: none;
+}
+
+.profile-picture {
+    width: 48px;
+    height: 48px;
+    flex-shrink: 0;
+    overflow: hidden;
+    border-radius: 50%;
+    margin-right: 20px;
+    display: flex;
+    background-color: white;
+}
+
+.profile-picture img {
+    max-width: 100%;
+    max-height: 100%;
+    object-fit: contain;
+}
+
+.profile-popup {
+    margin-right: 16px;
+}
+
+
+.sidebar {
+    background: linear-gradient(to bottom, #141414, #252525) fixed;
+    position: fixed;
+    height: 100%;
+    width: 92px;
+    display: flex;
+    flex-direction: column;
+    justify-content: end;
+}
+
+.sidebar__button-container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-self: end;
+    gap: 32px;
+    width: 100%;
+    margin-bottom: 36px;
+}
+
+.profile-container {
+    display: flex;
+}
+
+
+.scrollable-area {
+    overflow-y: auto;
+    overflow-x: clip;
+    width: 100%;
+    max-height: calc(100vh - 112px - 36px - 36px);
+}
+
+.scrollable-area::-webkit-scrollbar {
+    width: 10px;
+    border-radius: 10px;
+}
+
+.scrollable-area::-webkit-scrollbar-track {
+    background: linear-gradient(to bottom, #141414, #252525) fixed;
+}
+
+.scrollable-area::-webkit-scrollbar-thumb {
+    background: #888;
+    border-radius: 5px;
+}
+
+.scrollable-area::-webkit-scrollbar-thumb:hover {
+    background: #555;
+}
