+++ public/css/feed.css
@@ -1,81 +1,3 @@
-.index-container {
-    display: grid;
-    grid-template-columns: 80px 1fr;
-    grid-template-rows: 112px 1fr;
-    background: linear-gradient(to bottom, #141414, #252525) fixed;
-}
-
-
-.navbar {
-    position: fixed;
-    background: linear-gradient(to bottom, #141414, #252525) fixed;
-    display: flex;
-    width: 100%;
-    height: 112px;
-    padding: 32px 0;
-    align-items: center;
-    justify-content: space-between;
-    z-index: 100;
-}
-
-.navbar-logo {
-    align-self: center;
-    margin: 0 10px;
-    float: left;
-}
-
-.search-form {
-    flex-grow: 1;
-    height: 48px;
-    margin-right: 20px;
-    padding: 10px;
-    border-radius: 20px;
-    border: none;
-}
-
-.profile-picture {
-    width: 48px;
-    height: 48px;
-    flex-shrink: 0;
-    overflow: hidden;
-    border-radius: 50%;
-    margin-right: 20px;
-    display: flex;
-    background-color: white;
-}
-
-.profile-picture img {
-    max-width: 100%;
-    max-height: 100%;
-    object-fit: contain;
-}
-
-.profile-popup {
-    margin-right: 16px;
-}
-
-
-.sidebar {
-    background: linear-gradient(to bottom, #141414, #252525) fixed;
-    position: fixed;
-    height: 100%;
-    width: 80px;
-    display: flex;
-    flex-direction: column;
-    justify-content: end;
-}
-
-.sidebar__button-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-self: end;
-    gap: 32px;
-    width: 100%;
-    margin-bottom: 36px;
-}
-
-
 .feed {
     grid-row-start: 2;
     grid-column-start: 2;
@@ -90,7 +12,6 @@
     border-radius: 20px;
     background-color: #303030;
     margin-bottom: 32px;
-    /*margin-right: 32px;*/
 }
 
 .skeleton-1 {
