+++ public/index.html
@@ -12,6 +12,6 @@
 </head>
 <body>
     <div id="root"></div>
-    <script src="/src/index.js" type="module"></script>
+    <script src="/index.js" type="module"></script>
 </body>
 </html>
