+++ static/html/feed.html
@@ -24,13 +24,13 @@
         <div class="sidebar">
             <div class="sidebar__button-container">
                 <a id="newPin">
-                    <img src="../../public/icons/new-pin-icon.svg" alt="new pin">
+                    <img src="../../public/icons/new-pin.svg" alt="new pin">
                 </a>
                 <a id="chats">
-                    <img src="../../public/icons/chat-icon.svg" alt="chats">
+                    <img src="../../public/icons/chat.svg" alt="chats">
                 </a>
                 <a id="logout">
-                    <img src="../../public/icons/settings-icon.svg" alt="settings">
+                    <img src="../../public/icons/log-out.svg" alt="settings">
                 </a>
             </div>
         </div>
