+++ static/html/login.html
@@ -23,14 +23,14 @@
                 <label for="email" class="input-field-label">Email</label>
                 <div class="input-container">
                     <input type="text" class="input-container__field" id="email" placeholder="Введите email">
-                    <img src="../../public/icons/error-icon.svg" class="input-container__error hidden" aria-hidden="true" alt="">
+                    <img src="../../public/icons/error.svg" class="input-container__error hidden" aria-hidden="true" alt="">
                 </div>
                 <p class="error-message error-label hidden" id="email-error">Неверный формат email</p>
 
                 <label for="password" class="input-field-label">Пароль</label>
                 <div class="input-container">
                     <input type="password" class="input-container__field  input-container--invalid" id="password" placeholder="Введите пароль">
-                    <img src="../../public/icons/error-icon.svg" class="input-container__error" aria-hidden="true" alt="">
+                    <img src="../../public/icons/error.svg" class="input-container__error" aria-hidden="true" alt="">
                 </div>
                 <p class="error-message error-label" id="password-error">Неверный пароль или почта</p>
             </div>
