+++ static/html/one-pin.html
@@ -1,0 +1,99 @@
+<!DOCTYPE html>
+<html lang="ru">
+<head>
+  <meta charset="UTF-8">
+  <link rel="stylesheet" href="/css/fonts.css">
+  <link rel="stylesheet" href="/css/common.css">
+  <link rel="stylesheet" href="/css/one-pin.css">
+  <title>Flow</title>
+</head>
+
+<body>
+  <div class="main-page">
+    <div class="navbar">
+      <img src="../../public/icons/logo-small.svg" class="navbar-logo" alt="site logo">
+      <input type="text" class="search-form" placeholder="Text for search..." id="search">
+      <div class="profile-picture">
+        <img src="../../public/img/pfp1.jpg" alt="profile picture">
+      </div>
+      <a href="#" class="profile-popup">
+        <img src="../../public/icons/arrow-down.svg" aria-hidden="true" alt="">
+      </a>
+    </div>
+
+    <div class="sidebar">
+      <div class="sidebar__button-container">
+        <a id="newPin">
+          <img src="../../public/icons/new-pin.svg" alt="new pin">
+        </a>
+        <a id="chats">
+          <img src="../../public/icons/chat.svg" alt="chats">
+        </a>
+        <a id="logout">
+          <img src="../../public/icons/settings-icon.svg" alt="settings">
+        </a>
+      </div>
+    </div>
+
+    <div class="post">
+      <div class="post-picture">
+        <img src="/img/test-picture.jpg" class="flex-image" alt="">
+      </div>
+
+      <div class="one-pin">
+        <div class="menu">
+          <div class="button-container">
+            <img src="/icons/heart.svg" class="menu-button" id="like" alt="like">
+            <p class="input-text counter">5</p>
+            <img src="/icons/chat.svg" class="menu-button" id="newComment" alt="comments">
+            <p class="input-text counter">7</p>
+          </div>
+          <div class="back-to-feed">
+            <img src="/icons/arrow-left.svg" class="menu-button" id="gotoFeed" alt="back to feed">
+            <p>обратно к flows</p>
+          </div>
+        </div>
+
+        <div class="scrollable-area">
+          <div class="pin-text-field">
+            <h1 class="header" id="pin-name">Pin name</h1>
+            <div class="pin-text">
+              I'm blue, da-ba-dee, da-ba-di
+              <br>Da-ba-dee, da-ba-di, da-ba-dee, da-ba-di
+              <br>Da-ba-dee, da-ba-di, da-ba-dee, da-ba-di
+              <br>Da-ba-dee, da-ba-di, da-ba-dee, da-ba-di
+              <br>I'm blue, da-ba-dee, da-ba-di
+              <br>Da-ba-dee, da-ba-di, da-ba-dee, da-ba-di
+              <br>Da-ba-dee, da-ba-di, da-ba-dee, da-ba-di
+              <br>Da-ba-dee, da-ba-di, da-ba-dee, da-ba-di
+            </div>
+          </div>
+
+          <div class="author-container">
+            <img src="../../public/img/pfp1.jpg" class="profile-picture" alt="profile picture">
+            <p class="nickname">Jack Fresko</p>
+          </div>
+
+          Комментарии
+
+          <div class="commentary">
+            <div class="profile-container">
+              <img src="../../public/img/pfp1.jpg" class="profile-picture" alt="profile picture">
+              <p class="nickname">Jack Fresko</p>
+            </div>
+            <div class="commentary-text">
+              I'm blue, da-ba-dee, da-ba-di
+              <br>Da-ba-dee, da-ba-di, da-ba-dee, da-ba-di
+              <br>Da-ba-dee, da-ba-di, da-ba-dee, da-ba-di
+              <br>Da-ba-dee, da-ba-di, da-ba-dee, da-ba-di
+              <br>I'm blue, da-ba-dee, da-ba-di
+              <br>Da-ba-dee, da-ba-di, da-ba-dee, da-ba-di
+              <br>Da-ba-dee, da-ba-di, da-ba-dee, da-ba-di
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+</body>
+</html>
