+++ static/html/signup.html
@@ -22,35 +22,35 @@
                 <label for="email" class="input-field-label">Email</label>
                 <div class="input-container">
                     <input type="text" class="input-container__field" id="email" placeholder="">
-                    <img src="../../public/icons/error-icon.svg" class="input-container__error hidden" aria-hidden="true" alt="">
+                    <img src="../../public/icons/error.svg" class="input-container__error hidden" aria-hidden="true" alt="">
                 </div>
                 <p class="error-message error-label hidden" id="email-error">Неверный формат email</p>
 
                 <label for="nickname" class="input-field-label">Имя пользователя</label>
                 <div class="input-container">
                     <input type="text" class="input-container__field" id="nickname" placeholder="">
-                    <img src="../../public/icons/error-icon.svg" class="input-container__error hidden" aria-hidden="true" alt="">
+                    <img src="../../public/icons/error.svg" class="input-container__error hidden" aria-hidden="true" alt="">
                 </div>
                 <p class="error-message error-label hidden" id="nickname-error">Это имя уже занято</p>
 
                 <label for="birthday" class="input-field-label">Дата рождения</label>
                 <div class="input-container">
                     <input type="date" class="input-container__field input-container--invalid" id="birthday">
-                    <img src="../../public/icons/error-icon.svg" class="input-container__error" aria-hidden="true" alt="">
+                    <img src="../../public/icons/error.svg" class="input-container__error" aria-hidden="true" alt="">
                 </div>
                 <p class="error-message error-label" id="birthday-error">Некорректная дата рождения</p>
 
                 <label for="password">Пароль</label>
                 <div class="input-container">
                     <input type="password" class="input-container__field" id="password" placeholder="">
-                    <img src="../../public/icons/error-icon.svg" class="input-container__error hidden" aria-hidden="true" alt="">
+                    <img src="../../public/icons/error.svg" class="input-container__error hidden" aria-hidden="true" alt="">
                 </div>
                 <p class="error-message error-label hidden" id="password-error">Пароль слишком простой</p>
 
                 <label for="password-confirm" class="input-field-label">Пароль ещё раз</label>
                 <div class="input-container">
                     <input type="password" class="input-container__field" id="password-confirm" placeholder="">
-                    <img src="../../public/icons/error-icon.svg" class="input-container__error hidden" aria-hidden="true" alt="">
+                    <img src="../../public/icons/error.svg" class="input-container__error hidden" aria-hidden="true" alt="">
                 </div>
                 <p class="error-message error-label hidden" id="password-confirm-error">Пароли не совпадают</p>
             </form>
